Add vim-lastplace
It's awesome, because it just works. There is no thinking about how to use this plugin. Just install it, and boom, when you open a file it opens exactly where you left off last time. There are many other ways of doing this on the web, but none work as well as this one and none are as carefully thought out. It's configurable so that it knows not to jump to the last edit position for specific file types such as commit messages. This feature ensures that the plugin never annoys and always delights!