Add Surround-Funk plugin
This plugin was inspired by tpope's surround.vim and allows you to delete, change and yank a surrounding function along with its additional arguments. With the surrounding function in the unnamed register, you can 'grip' a word or another function with it. 'Gripping' will wrap/encompass a word or function with the one you have in the unnamed register.